start on bug fix for bad artifacts of zonal extraction - Untangling the puzzle
I'm confused as well, but I have re-extracted the data using the new
method and this results in the exact same result. Final value of any
duplicate is the correct one, but with the disputed areas it is only
where there are 3 duplicates it is correct. All disputed areas were
assigned 3 times to 1 country and 2 times to the other country. The 3rd
value in the country it is assigned 3 times to is the correct one. For
the other boundaries where the admin name was carried across - it is the
second of the 2 values. I'm not going to pretend I know why this works,
but this seems to be the pattern after a lot of trial and error.