Fix zonal extraction to use gaul codes to avoid matching error with multi-polygons from dispute areas
Also added the start of some code to agggregate up from admin 2
boundaries, but for it to work for the average correctly we need the
number of cells per admin area and weight each mean by that. I have a
feeling this will take as long or longer than just running the
extractions on all 3 admin levels and greatly increase potential for
error. The logic will be useful to combine disputed areas into their
countries for the mean, and the statistics required to do that have not
been calculated