Start of fixes for 2.1 - Still not working due to issue in the find_consecutive_patterns results.
Some sorting or similar needs to be applied by month and year. Lots of
other fixes related to gaul0_codes and other variables that were
missing/renamed. I also hashed out some of the baseline info in the data
json descriptors as it was causing breaks and the data fixes were a
higher priority